+++ .env
@@ -1,1 +1,1 @@
-VUE_APP_GRAPHQL_HTTP=https://graphqlzero.almansi.me/api
+VUE_APP_GRAPHQL_HTTP=https://tender-backend-dev.azurewebsites.net/graphql
