+++ src/graphql/User.gql
@@ -1,0 +1,5 @@
+query User {
+    generateUser {
+        uuid
+    }
+}
