+++ src/main.js
@@ -16,8 +16,8 @@ Vue.component('font-awesome-icon', FontAwesomeIcon)
 Vue.config.productionTip = false
 
 new Vue({
+  apolloProvider: createProvider(),
   router,
   store,
-  apolloProvider: createProvider(),
   render: h => h(App)
 }).$mount('#app')
