--- src/store/user.js
@@ -1,17 +1,0 @@
-const userStorage = {
-    state: () =>({
-        isAuthenticated: false,
-        authenticationToken: localStorage.getItem("tender-user-token") || "",
-    }),
-    mutations: {},
-    actions: {
-        async authRequest(context) {
-
-        }
-    },
-    getters: {},
-}
-
-export default {
-    userStorage
-}
